Перенесены функции поиска.
